requirements.txt
fastapi
uvicorn[standard]
python-multipart
jinja2
torch
torchvision
transformers
segmentation-models-pytorch
albumentations
opencv-python-headless
numpy
Pillow
